refactor json object entry ops
